chore: revert tag controller
