+++ packages/consumption/package.json
@@ -59,7 +59,6 @@
         }
     },
     "dependencies": {
-        "@js-soft/docdb-access-abstractions": "1.1.0",
         "@js-soft/docdb-querytranslator": "^1.1.5",
         "@js-soft/ts-serval": "2.0.12",
         "@js-soft/ts-utils": "2.3.3",
