+++ packages/consumption/src/modules/index.ts
@@ -6,4 +6,3 @@ export * from "./identityMetadata";
 export * from "./notifications";
 export * from "./requests";
 export * from "./settings";
-export * from "./tags";
