--- packages/consumption/src/modules/tags/index.ts
@@ -1,1 +1,0 @@
-export * from "./TagsController";
