+++ packages/consumption/test/modules/attributes/AttributesController.test.ts
@@ -1,4 +1,5 @@
 import { IDatabaseConnection } from "@js-soft/docdb-access-abstractions";
+import { sleep } from "@js-soft/ts-utils";
 import {
     BirthDate,
     BirthYear,
@@ -20,10 +21,13 @@ import {
     ZipCode
 } from "@nmshd/content";
 import { CoreAddress, CoreDate, CoreId } from "@nmshd/core-types";
-import { AccountController, Transport } from "@nmshd/transport";
+import { AccountController, ClientResult, TagClient, Transport } from "@nmshd/transport";
+import { anything, reset, spy, verify, when } from "ts-mockito";
 import {
     AttributeCreatedEvent,
     AttributeDeletedEvent,
+    AttributesController,
+    AttributeTagCollection,
     ConsumptionController,
     IAttributeSuccessorParams,
     ICreateRepositoryAttributeParams,
@@ -3368,5 +3372,130 @@ describe("AttributesController", function () {
         const peerAttribute = await consumptionController.attributes.getLocalAttribute(peerRelationshipAttribute.id);
         expect(ownAttribute).toBeUndefined();
         expect(peerAttribute).toBeUndefined();
+    });
+
+    describe("tag definition caching by time", function () {
+        let connection: IDatabaseConnection;
+        let transport: Transport;
+        let testAccount: AccountController;
+        let consumptionController: ConsumptionController;
+        let tagClientSpy: TagClient;
+        beforeEach(async function () {
+            connection = await TestUtil.createConnection();
+            transport = TestUtil.createTransport(connection, mockEventBus, {
+                tagCacheLifetimeInMinutes: 1 / 60
+            });
+            await transport.init();
+
+            const connectorAccount = (await TestUtil.provideAccounts(transport, 1))[0];
+            ({ accountController: testAccount, consumptionController } = connectorAccount);
+            const tagClient = consumptionController.attributes["attributeTagClient"];
+
+            tagClientSpy = spy(tagClient);
+            when(tagClientSpy.get("/api/v1/Tags", anything(), anything())).thenResolve(
+                ClientResult.ok(
+                    AttributeTagCollection.from({
+                        supportedLanguages: ["en"],
+                        tagsForAttributeValueTypes: {}
+                    }).toJSON(),
+                    {
+                        etag: "some-e-tag"
+                    }
+                )
+            );
+        });
+
+        afterEach(async function () {
+            await testAccount.close();
+            await connection.close();
+            reset(tagClientSpy);
+        });
+
+        test("should cache the tag definitions when called twice within tagCachingDurationInMinutes", async function () {
+            await consumptionController.attributes.getAttributeTagCollection();
+            await consumptionController.attributes.getAttributeTagCollection();
+
+            verify(tagClientSpy.getTagCollection(anything())).once();
+            reset(tagClientSpy);
+        });
+
+        test("should not cache the tag definitions when the tagCachingDurationInMinutes was reached", async function () {
+            await consumptionController.attributes.getAttributeTagCollection();
+            await sleep(1100);
+            await consumptionController.attributes.getAttributeTagCollection();
+
+            verify(tagClientSpy.getTagCollection(anything())).twice();
+            reset(tagClientSpy);
+        });
+    });
+
+    describe("tag definition caching by e-tag", function () {
+        let connection: IDatabaseConnection;
+        let transport: Transport;
+        let testAccount: AccountController;
+        let consumptionController: ConsumptionController;
+        let tagClientSpy: TagClient;
+        let attributesControllerSpy: AttributesController;
+        let etag: string;
+        beforeEach(async function () {
+            connection = await TestUtil.createConnection();
+            transport = TestUtil.createTransport(connection, mockEventBus, {
+                tagCacheLifetimeInMinutes: 0
+            });
+            await transport.init();
+
+            const connectorAccount = (await TestUtil.provideAccounts(transport, 1))[0];
+            ({ accountController: testAccount, consumptionController } = connectorAccount);
+            const tagClient = consumptionController.attributes["attributeTagClient"];
+
+            tagClientSpy = spy(tagClient);
+            attributesControllerSpy = spy(consumptionController.attributes);
+
+            etag = "some-e-tag";
+            when(tagClientSpy.get("/api/v1/Tags", anything(), anything())).thenCall((_path, _params, config) => {
+                const etagMatched = etag === config?.headers?.["if-none-match"];
+                const platformParameters = {
+                    etag,
+                    responseStatus: etagMatched ? 304 : 200
+                };
+                return Promise.resolve(
+                    ClientResult.ok(
+                        etagMatched
+                            ? undefined
+                            : AttributeTagCollection.from({
+                                  supportedLanguages: ["en"],
+                                  tagsForAttributeValueTypes: {}
+                              }).toJSON(),
+                        platformParameters
+                    )
+                );
+            });
+        });
+
+        afterEach(async function () {
+            await testAccount.close();
+            await connection.close();
+            reset(tagClientSpy);
+            reset(attributesControllerSpy);
+        });
+
+        test("should cache the tag definitions when called twice without new etag", async function () {
+            await consumptionController.attributes.getAttributeTagCollection();
+            await sleep(100);
+            await consumptionController.attributes.getAttributeTagCollection();
+
+            verify(tagClientSpy.getTagCollection(anything())).twice();
+            verify(attributesControllerSpy["setTagCollection"](anything())).once();
+        });
+
+        test("should not cache the tag definitions when called twice with new etag", async function () {
+            await consumptionController.attributes.getAttributeTagCollection();
+            await sleep(100);
+            etag = "some-other-e-tag";
+            await consumptionController.attributes.getAttributeTagCollection();
+
+            verify(tagClientSpy.getTagCollection(anything())).twice();
+            verify(attributesControllerSpy["setTagCollection"](anything())).twice();
+        });
     });
 });
